+++ src/components/Simple/View.tsx
@@ -1,71 +1,76 @@
-import { useMemo } from 'react';
 import type { FlexAlignType, StyleProp, ViewProps as RnViewProps, ViewStyle } from 'react-native';
-import { View as RnView } from 'react-native';
+import { StyleSheet, View as RnView } from 'react-native';
 
 
 
-type JustifyContent = 'center' | 'flex-start' | 'flex-end' | 'space-between' | 'space-around' | 'space-evenly' | undefined;
-
-
-export type NewViewProps = {
+export type GevViewProps = {
   /** Alias to `style` */
   s?: StyleProp<ViewStyle>;
-  /** Shortcut to `style: {flexDirection: 'row'}` */
+  /** Shortcut to `style: { flexDirection: 'row' }`. */
   row?: boolean;
-  /** If shall use 'row-reverse' / 'column-reverse' instead of 'row' / 'column' */
+  /** If shall use 'row-reverse' / 'column-reverse' instead of 'row' / 'column'. */
   reverse?: boolean;
-  /** Shortcut to `style: {justifyContent: X}` */
-  justify?: JustifyContent;
-  /** Shortcut to `style: {alignItems: X}`.
-   *
-   * If **true**, defaults to `'center'`. */
+  /** Shortcut to `style: { justifyContent: X }`. */
+  justify?: ViewStyle['justifyContent'];
+  /** Shortcut to `style: { alignItems: X }`. `'center'` if true. */
   align?: true | FlexAlignType;
-  /** Centers children in main and cross axis.
-   *
-   * Shortcut to style `style: {justifyContent: 'center', alignItems: 'center'}` */
+  /** Shortcut to `style: { justifyContent: 'center', alignItems: 'center' }`. */
   center?: boolean;
-  /** Shortcut to `style: {flex: 1}`. You may also pass a number to be used instead of the default 1, when **true**. */
-  flex?: boolean | number;
+  /** Shortcut to `style: { flex: X }`. `1` if true. */
+  flex?: true | number;
+  /** Shortcut to `style: { flexShrink: X }`. `1` if true. */
+  shrink?: true | number;
+  /** Shortcut to `style: { flexGrow: X }`. `1` if true. */
+  grow?: true | number;
 };
 
-export interface ViewProps extends RnViewProps, NewViewProps {}
+export interface ViewProps extends RnViewProps, GevViewProps {}
 
 /** ViewProps with some problematic props like hitSlop omitted. Pressable hitSlop type def is different. */
 type MergeViewStylesProps = Omit<ViewProps, 'hitSlop'>;
 
+
 export function mergeViewStyles({
-  row, reverse, justify, align: alignProp, center, s, style, flex: flexProp,
+  row, reverse, justify, align, center, s, style,
+  flex, grow, shrink,
 }: MergeViewStylesProps): StyleProp<ViewStyle> {
-  const align = alignProp === true ? 'center' : alignProp;
-  const flex = flexProp === true ? 1 : (typeof flexProp === 'number' ? flexProp : undefined);
   return [{
-    flex,
+    flex: flex === true ? 1 : flex,
+    flexGrow: grow === true ? 1 : grow,
+    flexShrink: shrink === true ? 1 : shrink,
     flexDirection: (row
       ? (reverse ? 'row-reverse' : 'row')
       : (reverse ? 'column-reverse' : 'column')),
     justifyContent: justify,
-    alignItems: align,
-    ...center && { justifyContent: 'center', alignItems: 'center' },
+    alignItems: align === true ? 'center' : align,
+    ...center && styles.center,
   }, s, style];
 }
 
-/** Simple wrapper for View component. It adds the properties:
+/** Wrapper for the View component. It adds the properties:
  *
- * * `s`, alias to `style`
- * * `row`, shortcut for style: `{flexDirection: 'row'}`
+ * * `s`, alias to `style`.
+ * * `row`, shortcut for style: `{flexDirection: 'row'}`.
  * * `reversed`, to concatenate '-reversed' to 'row' or 'column' in flexDirection.
- * * `center`, shortcut to style `style: {justifyContent: 'center', alignItems: 'center'}`
- * * `justify`, shortcut to `style: {justifyContent: X}`
+ * * `center`, shortcut to style `style: {justifyContent: 'center', alignItems: 'center'}`.
+ * * `justify`, shortcut to `style: {justifyContent: X}`.
  * * `align`, shortcut to `style: {alignItems: X}`. If true, it defaults to `'center'`.
- * * `onPress`, to use a Pressable instead of a View!
+ * * `flex`, shortcut to `style: { flex: X }`. `1` if true.
+ * * `grow`, shortcut to `style: { flex: X }`. `1` if true.
+ * * `shrink`, shortcut to `style: { flex: X }`. `1` if true.
  * */
 export const View: React.FC<ViewProps> = (props) => {
-  const style = useMemo(() => mergeViewStyles(props), [props]);
-
   return (
     <RnView
       {...props}
-      style={style}
+      style={mergeViewStyles(props)}
     />
   );
 };
+
+const styles = StyleSheet.create({
+  center: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});